+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Log of changes in the versions
 ## v1.5.1.2
 
 - correctly testing vector qualifications
+- improve on description generation when constructing a standard name from qualifications
 
 ## v1.5.1.1